Fix trader capacity roll and add optional Enhanced 1600 trip cap.
Widen capacity_random to uint16, use inclusive caravan rolls, and gate buy/sell plus ship/docker limits via max_capacity with a spawn-time ship snapshot.

+++ src/figuretype/figure_kingdome_trader.h
@@ -31,7 +31,7 @@ public:
         uint16_t wait_ticks_after_create;
         uint16_t min_capacity;
         uint16_t max_capacity;
-        uint8_t capacity_random;
+        uint16_t capacity_random;
     } FIGURE_STATIC_DATA_T;
 
     virtual void on_create() override;
